requirements.txt
discord.py
pydantic
PyNaCl
python-dotenv